Resolved requested changes
