+++ nextjs-frontend/src/components/my-catalog/table-sorting-functions/descendingComparator.tsx
@@ -1,0 +1,9 @@
+export function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
+    if (b[orderBy] < a[orderBy]) {
+        return -1;
+    }
+    if (b[orderBy] > a[orderBy]) {
+        return 1;
+    }
+    return 0;
+}
