+++ nextjs-frontend/src/components/my-catalog/table-sorting-functions/getComparator.tsx
@@ -1,0 +1,15 @@
+import {descendingComparator} from "<components>/components/my-catalog/table-sorting-functions/descendingComparator";
+
+export type Order = 'asc' | 'desc';
+
+export function getComparator<Key extends keyof any>(
+    order: Order,
+    orderBy: Key,
+): (
+    a: { [key in Key]: number | string },
+    b: { [key in Key]: number | string },
+) => number {
+    return order === 'desc'
+        ? (a, b) => descendingComparator(a, b, orderBy)
+        : (a, b) => -descendingComparator(a, b, orderBy);
+}
