+++ nextjs-frontend/src/components/my-catalog/table-utils/EnhancedTableProps.tsx
@@ -1,0 +1,11 @@
+import * as React from "react";
+import {Order} from "<components>/components/my-catalog/table-sorting-functions/getComparator";
+
+export interface EnhancedTableProps {
+    numSelected: number;
+    onRequestSort: (event: React.MouseEvent<unknown>, property: keyof ProductData) => void;
+    onSelectAllClick: (event: React.ChangeEvent<HTMLInputElement>) => void;
+    order: Order;
+    orderBy: string;
+    rowCount: number;
+}
