+++ nextjs-frontend/src/components/my-catalog/table-utils/EnhancedTableToolbarProps.tsx
@@ -1,0 +1,3 @@
+interface EnhancedTableToolbarProps {
+    numSelected: number;
+}
