+++ nextjs-frontend/src/components/my-catalog/table-utils/HeadCell.tsx
@@ -1,0 +1,6 @@
+interface HeadCell {
+    disablePadding: boolean;
+    id: keyof ProductData;
+    label: string;
+    numeric: boolean;
+}
