+++ nextjs-frontend/src/components/my-catalog/table-utils/headCells.tsx
@@ -1,0 +1,44 @@
+export const headCells: readonly HeadCell[] = [
+    {
+        id: 'name',
+        numeric: false,
+        disablePadding: true,
+        label: 'Product Name',
+    },
+    {
+        id: 'brand',
+        numeric: true,
+        disablePadding: false,
+        label: 'Brand',
+    },
+    {
+        id: 'cost',
+        numeric: true,
+        disablePadding: false,
+        label: 'Cost',
+    },
+    {
+        id: 'price',
+        numeric: true,
+        disablePadding: false,
+        label: 'Price',
+    },
+    {
+        id: 'competitorPrices',
+        numeric: true,
+        disablePadding: false,
+        label: 'Competitor Prices',
+    },
+    {
+        id: 'position',
+        numeric: true,
+        disablePadding: false,
+        label: 'Position',
+    },
+    {
+        id: 'status',
+        numeric: true,
+        disablePadding: false,
+        label: 'Status',
+    },
+];
