+++ nextjs-frontend/src/model/ProductData.tsx
@@ -1,0 +1,10 @@
+interface ProductData {
+    name: string;
+    ean: number,
+    brand: string;
+    cost: number;
+    price: number;
+    competitorPrices: string;
+    position: number,
+    status: string
+}
