+++ nextjs-frontend/src/pages/dashboard/my-catalog/index.tsx
@@ -1,37 +1,22 @@
 import * as React from 'react';
 import DashboardDrawer from "<components>/components/dashboard-drawer/DashboardDrawer";
-import {alpha, Checkbox, Divider, Grid, TextField, Typography} from "@mui/material";
+import {Checkbox, Divider, Grid, TextField, Typography} from "@mui/material";
 import styles from '../../../styles/MyCatalog.module.css'
 import Box from '@mui/material/Box';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
 import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
 import TablePagination from '@mui/material/TablePagination';
 import TableRow from '@mui/material/TableRow';
-import TableSortLabel from '@mui/material/TableSortLabel';
-import Toolbar from '@mui/material/Toolbar';
 import Paper from '@mui/material/Paper';
-import IconButton from '@mui/material/IconButton';
-import Tooltip from '@mui/material/Tooltip';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import Switch from '@mui/material/Switch';
-import DeleteIcon from '@mui/icons-material/Delete';
-import AddBoxIcon from '@mui/icons-material/AddBox';
-import EditIcon from '@mui/icons-material/Edit';
-import {visuallyHidden} from '@mui/utils';
+import {getComparator, Order} from "<components>/components/my-catalog/table-sorting-functions/getComparator";
+import {stableSort} from "<components>/components/my-catalog/table-sorting-functions/stableSort";
+import {EnhancedTableToolbar} from "<components>/components/my-catalog/table-utils/EnhancedTableToolbar";
+import {EnhancedTableHead} from "<components>/components/my-catalog/table-utils/EnhancedTableHead";
 
-interface Data {
-    name: string;
-    ean: number,
-    brand: string;
-    cost: number;
-    price: number;
-    competitorPrices: string;
-    position: number,
-    status: string
-}
 
 function createData(
     name: string,
@@ -42,7 +27,7 @@ function createData(
     competitorPrices: string,
     position: number,
     status: string
-): Data {
+): ProductData {
     return {
         name,
         ean,
@@ -74,223 +59,9 @@ const rows = [
     createData('Motorola Moto G99 256 GB', 194253401194, 'Motorola', 499.00, 799.00, 'Show', 1, 'Active'),
 ];
 
-function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
-    if (b[orderBy] < a[orderBy]) {
-        return -1;
-    }
-    if (b[orderBy] > a[orderBy]) {
-        return 1;
-    }
-    return 0;
-}
-
-type Order = 'asc' | 'desc';
-
-function getComparator<Key extends keyof any>(
-    order: Order,
-    orderBy: Key,
-): (
-    a: { [key in Key]: number | string },
-    b: { [key in Key]: number | string },
-) => number {
-    return order === 'desc'
-        ? (a, b) => descendingComparator(a, b, orderBy)
-        : (a, b) => -descendingComparator(a, b, orderBy);
-}
-
-function stableSort<T>(array: readonly T[], comparator: (a: T, b: T) => number) {
-    const stabilizedThis = array.map((el, index) => [el, index] as [T, number]);
-    stabilizedThis.sort((a, b) => {
-        const order = comparator(a[0], b[0]);
-        if (order !== 0) {
-            return order;
-        }
-        return a[1] - b[1];
-    });
-    return stabilizedThis.map((el) => el[0]);
-}
-
-interface HeadCell {
-    disablePadding: boolean;
-    id: keyof Data;
-    label: string;
-    numeric: boolean;
-}
-
-const headCells: readonly HeadCell[] = [
-    {
-        id: 'name',
-        numeric: false,
-        disablePadding: true,
-        label: 'Product Name',
-    },
-    {
-        id: 'brand',
-        numeric: true,
-        disablePadding: false,
-        label: 'Brand',
-    },
-    {
-        id: 'cost',
-        numeric: true,
-        disablePadding: false,
-        label: 'Cost',
-    },
-    {
-        id: 'price',
-        numeric: true,
-        disablePadding: false,
-        label: 'Price',
-    },
-    {
-        id: 'competitorPrices',
-        numeric: true,
-        disablePadding: false,
-        label: 'Competitor Prices',
-    },
-    {
-        id: 'position',
-        numeric: true,
-        disablePadding: false,
-        label: 'Position',
-    },
-    {
-        id: 'status',
-        numeric: true,
-        disablePadding: false,
-        label: 'Status',
-    },
-];
-
-interface EnhancedTableProps {
-    numSelected: number;
-    onRequestSort: (event: React.MouseEvent<unknown>, property: keyof Data) => void;
-    onSelectAllClick: (event: React.ChangeEvent<HTMLInputElement>) => void;
-    order: Order;
-    orderBy: string;
-    rowCount: number;
-}
-
-function EnhancedTableHead(props: EnhancedTableProps) {
-    const {onSelectAllClick, order, orderBy, numSelected, rowCount, onRequestSort} =
-        props;
-    const createSortHandler =
-        (property: keyof Data) => (event: React.MouseEvent<unknown>) => {
-            onRequestSort(event, property);
-        };
-
-    return (
-        <TableHead>
-            <TableRow>
-                <TableCell padding="checkbox">
-                    <Checkbox
-                        color="primary"
-                        indeterminate={numSelected > 0 && numSelected < rowCount}
-                        checked={rowCount > 0 && numSelected === rowCount}
-                        onChange={onSelectAllClick}
-                        inputProps={{
-                            'aria-label': 'select all desserts',
-                        }}
-                    />
-                </TableCell>
-                {headCells.map((headCell) => (
-                    <TableCell
-                        key={headCell.id}
-                        align={headCell.numeric ? 'right' : 'left'}
-                        padding={headCell.disablePadding ? 'none' : 'normal'}
-                        sortDirection={orderBy === headCell.id ? order : false}
-                    >
-                        <TableSortLabel
-                            active={orderBy === headCell.id}
-                            direction={orderBy === headCell.id ? order : 'asc'}
-                            onClick={createSortHandler(headCell.id)}
-                        >
-                            {headCell.label}
-                            {orderBy === headCell.id ? (
-                                <Box component="span" sx={visuallyHidden}>
-                                    {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
-                                </Box>
-                            ) : null}
-                        </TableSortLabel>
-                    </TableCell>
-                ))}
-            </TableRow>
-        </TableHead>
-    );
-}
-
-interface EnhancedTableToolbarProps {
-    numSelected: number;
-}
-
-function EnhancedTableToolbar(props: EnhancedTableToolbarProps) {
-    const {numSelected} = props;
-
-    return (
-        <Toolbar
-            sx={{
-                pl: {sm: 2},
-                pr: {xs: 1, sm: 1},
-                ...(numSelected > 0 && {
-                    bgcolor: (theme) =>
-                        alpha(theme.palette.primary.main, theme.palette.action.activatedOpacity),
-                }),
-            }}
-        >
-            {numSelected > 0 ? (
-                <Typography
-                    sx={{flex: '1 1 100%'}}
-                    color="inherit"
-                    variant="subtitle1"
-                    component="div"
-                >
-                    {numSelected} selected
-                </Typography>
-            ) : (
-                <Typography
-                    sx={{flex: '1 1 100%'}}
-                    variant="h6"
-                    id="tableTitle"
-                    component="div"
-                >
-                    Products
-                </Typography>
-            )}
-            {numSelected === 1 ? (
-                <>
-                    <Tooltip title="Edit Product">
-                        <IconButton size="large">
-                            <EditIcon/>
-                        </IconButton>
-                    </Tooltip>
-                    <Tooltip title="Delete Product">
-                        <IconButton size="large">
-                            <DeleteIcon/>
-                        </IconButton>
-                    </Tooltip>
-                </>
-            ) : (
-                numSelected > 1 ? (
-                    <Tooltip title="Delete Selected Products">
-                        <IconButton size="large">
-                            <DeleteIcon/>
-                        </IconButton>
-                    </Tooltip>
-                ) : (
-                    <Tooltip title="Add Product">
-                        <IconButton size="large">
-                            <AddBoxIcon/>
-                        </IconButton>
-                    </Tooltip>
-                )
-            )}
-        </Toolbar>
-    );
-}
-
 export default function MyCatalog() {
     const [order, setOrder] = React.useState<Order>('asc');
-    const [orderBy, setOrderBy] = React.useState<keyof Data>('cost');
+    const [orderBy, setOrderBy] = React.useState<keyof ProductData>('cost');
     const [selected, setSelected] = React.useState<readonly string[]>([]);
     const [page, setPage] = React.useState(0);
     const [dense, setDense] = React.useState(false);
@@ -298,7 +69,7 @@ export default function MyCatalog() {
 
     const handleRequestSort = (
         event: React.MouseEvent<unknown>,
-        property: keyof Data,
+        property: keyof ProductData,
     ) => {
         const isAsc = orderBy === property && order === 'asc';
         setOrder(isAsc ? 'desc' : 'asc');
